ADD explicit casting

+++ tangermeme/tools/tomtom.py
@@ -227,7 +227,7 @@ def _p_values(gamma, B_cdfs, rr_inv, T_lens, nq, offset, n_bins, results):
 		results[i, 1] = 0
 		
 		for k in range(-nq + 1, nt):
-			score = 0.0
+			score = 0
 			overlap = 0
 			
 			if k < 0:
@@ -236,12 +236,12 @@ def _p_values(gamma, B_cdfs, rr_inv, T_lens, nq, offset, n_bins, results):
 						break
 
 					j, t_idx = uint64(j), uint64(total_offset + j + k)
-					score += uint64(gamma[j, rr_inv[t_idx]])
+					score += gamma[j, rr_inv[t_idx]]
 					overlap += 1
 			else:
 				for j in range(min(nq, nt-k)):                    
 					j, t_idx = uint64(j), uint64(total_offset + j + k)
-					score += uint64(gamma[j, rr_inv[t_idx]])
+					score += gamma[j, rr_inv[t_idx]]
 					overlap += 1
 		
 			score = score + (nq - overlap) * offset
@@ -249,7 +249,7 @@ def _p_values(gamma, B_cdfs, rr_inv, T_lens, nq, offset, n_bins, results):
 				if score == results[i, 1] and results[i, 2] >= overlap:
 					continue
 
-				results[i, 0] = B_cdfs[nt, uint64(score)-1]
+				results[i, 0] = B_cdfs[nt, uint64(score-1)]
 				results[i, 1] = score
 				results[i, 2] = k
 				results[i, 3] = overlap
